+++ src/Commands/Graph/RemoveAvailableSiteClassification.cs
@@ -1,5 +1,4 @@
-﻿
-using PnP.PowerShell.Commands.Attributes;
+﻿using PnP.PowerShell.Commands.Attributes;
 using PnP.PowerShell.Commands.Base;
 using System;
 using System.Collections.Generic;
@@ -8,12 +7,12 @@ using System.Management.Automation;
 
 namespace PnP.PowerShell.Commands.Graph
 {
-    [Cmdlet(VerbsCommon.Remove, "PnPSiteClassification")]
+    [Cmdlet(VerbsCommon.Remove, "PnPAvailableSiteClassification")]
     [RequiredMinimalApiPermissions("Directory.ReadWrite.All")]
+    [Alias("Remove-PnPSiteClassitication")]
     [OutputType(typeof(void))]
     public class RemoveSiteClassification : PnPGraphCmdlet
     {
-
         [Parameter(Mandatory = true)]
         public List<string> Classifications;