+++ src/Commands/Utilities/REST/RestHelper.cs
@@ -5,7 +5,6 @@ using System.Net;
 using System.Net.Http;
 using System.Text;
 using System.Text.Json;
-using System.Threading;
 using System.Threading.Tasks;
 using Microsoft.SharePoint.Client;